Link d_cursor_mng for all Wii versions

+++ src/d/d_com_inf_game.cpp
@@ -134,7 +134,7 @@ void dComIfG_play_c::setNowVibration(u8 i_vibration) {
     mItemInfo.mNowVibration = i_vibration;
 }
 
-u32 dComIfG_play_c::getNowVibration() {
+u8 dComIfG_play_c::getNowVibration() {
     return mItemInfo.mNowVibration;
 }
